Use PSR4 for tests
